+++ cpp/src/parquet/metadata.cc
@@ -35,6 +35,8 @@
 #include "parquet/schema_internal.h"
 #include "parquet/thrift_internal.h"
 
+#include <syslog.h>
+
 namespace parquet {
 
 const ApplicationVersion& ApplicationVersion::PARQUET_251_FIXED_VERSION() {
@@ -592,7 +594,7 @@ class FileMetaData::FileMetaDataImpl {
   FileMetaDataImpl() = default;
   
   ~FileMetaDataImpl() {
-    syslog(0, "FileMetaDataImpl Destructor: key_value use count: %ld, file_decrypter use count: %ld", key_value_metadata_.use_count(), file_decryptor_.use_count());
+    syslog(0, "FileMetaDataImpl Destructor: key_value use count: %d, file_decrypter use count: %d", (int) key_value_metadata_.use_count(), (int) file_decryptor_.use_count());
   }
 
   explicit FileMetaDataImpl(
